pyproject.toml
[project]
name = "opac-devops"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.75",
    "psycopg2-binary>=2.9.11",
]
